se agrega documento de css, el enrutador se ubica en su propio archivo, se renombran funciones
